refactor: move module into shapetastic package
